rearranging; em1d passing now
