+++ geoapps/inversion/electromagnetics/application.py
@@ -29,7 +29,7 @@ from geoapps import assets_path
 from geoapps.base.application import BaseApplication
 from geoapps.base.plot import PlotSelection2D
 from geoapps.base.selection import LineOptions, ObjectDataSelection, TopographyOptions
-from geoapps.inversion.utils import preprocess_data
+from geoapps.inversion.components.preprocessing import preprocess_data
 from geoapps.utils import geophysical_systems, warn_module_not_found
 from geoapps.utils.list import find_value
 from geoapps.utils.string import string_2_list
