+++ geoapps/utils/surveys.py
@@ -8,6 +8,8 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
+from geoapps.inversion.utils import get_unique_locations
+
 if TYPE_CHECKING:
     from geoapps.inversion.components.data import InversionData
 
@@ -19,7 +21,6 @@ from geoh5py.data import FloatData
 from geoh5py.objects import CurrentElectrode, PotentialElectrode
 from geoh5py.workspace import Workspace
 from scipy.spatial import cKDTree
-from SimPEG.survey import BaseSurvey
 
 from geoapps.utils.statistics import is_outlier
 
@@ -48,24 +49,6 @@ def get_containing_cells(
         raise TypeError("Mesh must be 'TreeMesh' or 'TensorMesh'")
 
     return inds
-
-
-def get_unique_locations(survey: BaseSurvey) -> np.ndarray:
-    if survey.source_list:
-        locations = []
-        for source in survey.source_list:
-            source_location = source.location
-            if source_location is not None:
-                if not isinstance(source_location, list):
-                    locations += [[source_location]]
-                else:
-                    locations += [source_location]
-            locations += [receiver.locations for receiver in source.receiver_list]
-        locations = np.vstack([np.vstack(np.atleast_2d(*locs)) for locs in locations])
-    else:
-        locations = survey.receiver_locations
-
-    return np.unique(locations, axis=0)
 
 
 def new_neighbors(distances: np.ndarray, neighbors: np.ndarray, nodes: list[int]):
